refactor some more
Extract all composer handling code from php_execute into its own module.
