+++ agent/fw_hooks.h
@@ -56,6 +56,7 @@ extern void nr_phpunit_enable(TSRMLS_D);
 extern void nr_predis_enable(TSRMLS_D);
 extern void nr_zend_http_enable(TSRMLS_D);
 extern void nr_monolog_enable(TSRMLS_D);
+extern void nr_composer_handle_autoload(const char* filename);
 
 /* Vulnerability Management Packages */
 extern void nr_drupal_version(void);
